Fix Terraform resource churn on API instances
Move EBS volume definition to a separate volume block. For the time
being we will specify this volume manually with an ID into a
variable. This isn't a long term solution, eventually we will move
the package data to an S3 bucket and not need this volume anymore.
For now, let's prevent future resource churn which occurs when we
manually mount volumes to our API instances

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>