[Builder] Run the development container in privileged mode.
This change is required for Studio builds to execute via the worker.

Adding a `--name` makes this container easier to resolve, meaning you
can run `docker exec -ti builder sh` to enter the container. Due to the
port mappings back to localhost, there can only be one Builder container
running anyway, so you may as well tack a nice name on it.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>