[hab] Refactor `hab plan init` Scaffolding detection.
This is a light refactoring with an eye to adding more in the future:

* Lexically sort the Scaffolding types in `match` arms, detection order,
and detection functions. This should help a future contributor to know
where to put things and should help maintainers when scanning through
the module.
* Drop extra scaffolding name matches other than the package identifier
and the canonical name (i.e. "ruby", "node", etc.). It is hard to tell
if some of the alternate scaffolding matching words might become full
fledged Scaffoldings in the future. To avoid collisions such as these,
it seems safer to have fewer matches. Additionally, with autodetection,
it's less likely that a user would call the explicit `--scaffolding`
option on the command line.
* Uppercase some words for consistency.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>