[sup] Install a package only if missing for `start` & `load`.
This resolves an issue in `hab-sup start` and `hab-sup load` where the
Supervisor was unconditionally attempting to install a package saving
the Service spec. Without an explicit update strategy, the Supervisor
should find an load the latest version of an installed package on disk.
Otherwise, newer releases of a package will be download and installed
whenever a new candidate is present in the Builder API.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>