[pkg-export-docker] Build in runtime ident for Docker package.
This change will build the fully qualified package identifier for the
version of `core/docker` which it was built against into the binary when
building under Habitat. If it is built in development, a package
identifier of `core/docker` will be used and the program will fail if
the package is not installed.

The package-resolving code was originally in `builder-worker`, but this
change extracted `resolve_cmd_in_pkg()` to `core`.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>