Allow plan-build to build composite packages
A "composite" package is a collection of services that run together on
a single Habitat supervisor, allowing users to implement "sidecar"
patterns with Habitat services.

A composite package *is* a Habitat package, but one that is purely
metadata. It is also generated from a `plan.sh` file, just like any
other Habitat package. Here is a minimal example:

```sh
pkg_origin="mystuff"
pkg_name="api"
pkg_type="composite"
pkg_version="1.0.0"

pkg_services=(
    mystuff/app-server
    mystuff/api-proxy
)
pkg_bind_map=(
    [mystuff/api-proxy]="http:mystuff/app-server"
)
```

This would create a `mystuff/api` package that bundles together an
application server and a proxy, and wires the `http` bind of the proxy
to the application server. When you install this package, both the
`mystuff/app-server` and `mystuff/api-proxy` services would be
downloaded, installed, and bound together properly.

To create a new kind of Habitat package, some new metadata files had
to be introduced. They are as follows:

* `TYPE`: Indicates whether the package is a "composite" or a
  "standalone" (i.e., every other Habitat package in existence). This
  is backward-compatible; if no `TYPE` file is present, a value of
  "standalone" is assumed.

  This file is now added to all packages, though.

The following metadata files exist only in composite packages.

* `SERVICES`: defines the services that are members of the
  composite. This file contains package identifiers for each service,
  but in contrast to other "dependency"-type metadata files, this file
  does *not* need to contain fully-qualified identifiers. When a
  composite package is installed, spec files are generated for each
  service; the `ident` field of that spec comes from this file. Thus,
  if you wanted to create a composite whose services could update to
  the latest stable release, you could provide simple
  `origin/name`-style idents. However, if you wanted to restrict
  services to the latest of a specific version, you could give
  `origin/name/version`-style idents.
* `RESOLVED_SERVICES`: When the composite is built, binds are
  validated against concrete releases of the services. Their
  fully-qualified identifiers are recorded here. When you install a
  composite package, these are the exact versions of the services that
  will be installed. They can, of course, update and evolve as time
  goes on (and according to your update strategies), but they provide
  a known starting point.
* `BIND_MAP`: often, services within a composite must be wired
  together in order to function properly. This file records which
  service satisfies the bind of another service within the
  composite. These are all validated at build time to be mutually
  consistent (e.g., you can't say Service A satisfies bind "X" of
  Service B if Service A does not export the necessary values). When a
  composite is installed, the binds for each service are constructed
  automatically from these values.

In the course of this work, the Bash script that makes up the
`plan-build` package was decomposed into separate files, with many
functions being refactored to be more consistent and easier to reason
about.

Signed-off-by: Christopher Maier <cmaier@chef.io>