Install and otherwise manipulate composite packages
In order to deal with composite packages, several new concepts were
introduced.

Composite Specs
===============

In order to track information about currently-installed composite
packages (needed to know which services are part of a composite in
order to act on them as a group, for instance), a "composite spec",
analogous to existing service specs, was introduced. This is a TOML
file that resides in `/hab/sup/default/composites`, and records the
specific release of the composite that is active, as well as the
identifier that it was *loaded* as. By way of explanation, consider
that when you run `hab pkg load core/redis`, you will download a
specific release of `core/redis`, but you loaded it as simply
`core/redis`, which will affect its update behavior, but also the
behavior of `hab svc load --force`. (This is perhaps only required for
as long as we interact with services through package identifiers,
which can be ambiguous, rather than service groups).

The specific release is used to determine which services are
considered part of the composite, as well as to retrieve bind
information from.

Services
========

Services now record which composite (if any) they are a member
of. This is used in `hab svc status` to give more context to the user.

Metadata
========

Composite packages have several new metadata files that need to be
consulted; `PackageArchive` and `PackageInstall` can now read these.

`hab` Command Line Client
=========================

This is the site of the most significant changes.

* `hab pkg install`

Installing a composite means installing every service in the
composite. The specific versions that are installed are taken from the
`RESOLVED_SERVICES` metadata file, guaranteeing that you will have the
exact packages that were used for verification at composite
build-time. These packages are installed just like any other, using a
fully-qualified package identifier.

* `hab svc start`

Starting a composite package starts all the services in the composite.

* `hab svc stop`

Stopping a composite package stops all the services in the composite.

* `hab svc status`

The name of the composite a service is in is displayed in the
output. Standalone services are denoted by the string "standalone".

* `hab svc load`

Loading a composite will download and install all services if
needed. A composite spec file is written out, recording information
needed for interaction with composites via `hab`. For each service in
the composite, a service spec file is automatically generated as
follows:

- The `ident` used is taken from the `SERVICES` metadata file.

- The update strategy, Builder URL, update channel, application /
  environment, service group, and topology are currently shared by all
  members of the composite.

- Binds for each service are generated from the contents of the
  `BIND_MAP` metadata file. To support the usecase of services needing
  binds that are *not* satisfied by other members of the composite, we
  have adapted the existing `--bind` option to allow the user to
  describe these binds. Only when loading composites, you may specify a
  three-part bind:

      `${service-name}:${bind_name}:${service_group}`.

  This is effectively a "normal" bind, prefixed by the service in the
  composite it applies to. If you were to think of a composite as one
  unified service (instead of as the container for multiple services
  that it is), you could think of these 3-part binds as the binds of
  that one service. If composites were to become more of an actual
  run-time construct (as opposed to a packaging / install-time
  construct), that would be an interesting angle to pursue, though it
  would likely require more extensive changes to the core of the
  system.

  The binds are structured under the assumption that all services are
  in the same service group. Due to how "organization" is handled in
  the Supervisor (and current uncertainty as to its future in the
  codebase), binds do _not_ include any organization
  information. Thus, composites will not function properly on any
  Supervisor running with an organization set. It is not certain if
  anyone even runs Habitat this way, and in any case, the upcoming
  IPC-based communication refactoring of the supervisor will provide
  the right way to obtain the correct information.

- No values are written for the "config_from" or service password. If
  we wish to support the former, we should discuss the desired UX,
  because we would need provide per-service values (the approach taken
  for binds, described above, may be applicable. The latter is only
  applicable to Windows, and the work of this commit does not yet
  support Windows builds of composite packages (that may come with a
  rewrite of plan building in Rust).

* `hab svc load --force`

This is the most complex part of this work.

If you do not change the ident you refer to the composite as, things
are relatively straightforward. If any flags are supplied on the CLI,
those values get set for all the composite's services (subject to the
restrictions discussed for first-time loading). Binds are a little
interesting (but still logically consistent with standalone services)
because we don't change them unless the user supplies a value for the
`--bind` option, in which case, we rewrite the binds of *all* the
services (because these extra binds act like binds of the composite as
a whole).

If the ident you're referring to the composite as changes, however,
we're going to be changing the whole world. Different versions of a
composite can have very different services; some will be in common,
some will be removed, and some will be added. Any existing specs for
services not in the new composite will be unloaded. The old
configurations of services present in both the old and new composites
are discarded completely in favor of the new values.

The only hitch is that if you aren't "moving forward" with the
composite, you might not end up bringing up a consistent set of
services, since Habitat runs the latest version of a service that will
satisfy the identifier in the service spec. (This would not be a
problem in composites that contain services described by
fully-qualified identifiers, since those exact versions would always
be run. However, such services would never update.)

Summary
=======

There are undoubtedly subtleties to running and managing composite
packages as currently formulated. The functionality present here
should provide a suitable foundation to explore the concept,
though. Planned refactorings of the Supervisor's internals, and the
interaction between a Supervisor and the `hab` CLI client will remove
some of the current complexity around composites.

In the future, there may also be benefit to promoting composites to a
run-time construct, which would potentially make reasoning about them
easier. This would require much more invasive changes to the core of
the Supervisor to make the UX work cleanly and without surprises (in
contrast, the changes here are largely on the periphery of the system,
working as much as possible with what currently exists).

Signed-off-by: Christopher Maier <cmaier@chef.io>