Use GitHub App integration over OAuth App for authentication
This removes the need for user scopes and replaces the default dev
credentials for our applications with default credentials for our
dev GitHub App. User scopes are no longer required when using a GitHub
App, the permissions are instead based on what the user grants what
the App is asking for and what they make available.

Emails aren't guraranteed to be available to the GitHub App API so
we are optionally checking for it's value before setting it.
Additionally teams are handled differently and we will need to
do a bit of work to re-enable feature flagging on the back end.
We aren't currently using it so this shouldn't matter too much.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>