[studio] Only honor `.studiorc` in `hab studio enter`.
This change restricts where a provided `.studiorc` gets sourced to only
interactive Studios, i.e with `hab studio enter`.

The danger of sourcing such a file when performing an automated, clean
build via `hab pkg build` is that the build program can't account for
whatever changes to the Studio environment a `.studiorc` might perform.
Moreover, this feature was intended more for development use cases,
implying interactively entering the Studio. Restricting the use case for
.studiorc` also makes automated building safer by being more consistent,
and not prone to intended (or unintended) injection of an outside
environment.

HAB_STUDIO_NOSTUDIORC
---------------------

An additional environment variable is also introduced,
`HAB_STUDIO_NOSTUDIORC`, which disables the sourcing of a `.studiorc`
when interactively entering a Studio (via `hab studio enter`). Setting
this environment variable to a non-empty value sets this source-skipping
behavior.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>