Link key cache into Docker export root filesystem
When exporting a Docker container, we now link in the Habitat key
cache. This is mainly to support initial experiences where users are
trying Habitat out, but have not uploaded keys to Builder (for
whatever reason). If the keys are not present in the cache, Habitat
will attempt to retrieve the public key from Builder. This will fail
if the key is not in Builder (obviously!), which can be confusing to
new users.

As coded, there is a change in behavior here; previously, an exported
Docker image would contain all the public keys downloaded in the
course of generating the image. Since we now are symlinking the key
directory in (and unlinking after we've downloaded the packages), the
keys will no longer be present in the image. In practice, this shouldn't
cause any issues. If you are running a Habitat Docker container with
an update strategy of "none", there's no need for keys anyway, because
nothing is going to be downloaded from Builder. If you're running with
an update strategy, you're going to be downloading new versions of
your software from Builder, which inherently means keys will be
available and downloaded as needed.

Signed-off-by: Christopher Maier <cmaier@chef.io>