[studio] Attempt to import public origin keys into Studios.
This change will optimistically import the public origin key if found in
the external key cache. The public origin key is not strictly required
in order to build and sign a package, but there are some uses cases
where having the public origin key is convenient such as a user who is
building their first Habitat package locally, even before creating an
origin and uploading their key to Builder.

A failure to import the public origin key will show an informational
message but will not cause the Studio to fail or abort--this is best
effort and a nice-to-have.

References #3524
Closes #3495

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>