[git] Lexically sort `.gitignore` & add `./secrets/` directory to Git.
Just a housekeeping chore to sort the `.gitignore` as this file tends to
get hard to mange after a while and duplicate entries are very easy to
produce. Also, `.delivery/cli.toml` is dropped as that directory is no
longer under version control.

Additionally, this adds a small "gitkeeper" to leave developers with a
`.secrets/` directory when they clone this repo, making it a little
clearer where the GitHubApp secret should live.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>