Emit paths where the event happened
We will be running tests under a temporary directory (likely
/tmp/filewatcher-XXXXXX), so file watcher will watch both / and /tmp
and will receive events happening there.

In the tests, we are going to have steps, which are basically a file
system operation (like remove a file or create a symlink) and a
description of expected state after the operation. The operations will
happen entirely under the temporary directory. For each operation, we
compute how many times we need to iterate the watcher before checking
if the actual state is the same as the expected state.

Since we will be watching parent directories of the temporary
directory, events there may interfere with the computed number of
iterations. This is especially true for /tmp, where other processes
put their stuff.

The new callback can be used to work around the problem - the
implementation can check if the event happened somewhere in parent
directories of the temporary directory and "ignore" the iteration.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>