Add facilities for nicer debug output.
Using Debug traits on a struct that holds two hashmaps results in a
completely unreadable output. This commits adds a trait the types can
implement to provide a multiline and indented debug output, and
implementations of the trait for types used in the FileWatcher and its
tests.

This also adds a DebugInfo structure to store the most recent
debugging information about the executed test case. This is to get a
relatively short and relevant output in case of test failure instead
of all the information from the previous tests cases or steps.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>