Add more comprehensive tests for FileWatcher
Testing it is quite involved because beside testing if the outcome of
some operations is what we expect, we want to test some of the
internal tracking structures too.

The tests have basically two phases - an initial phase where we set up
the watcher with the initial path and we set up the initial
directories, files and symlinks, and a "steps" phase, where we perform
some file system action (move a directory, add a symlink, remove a
file, etc.) and observe the outcome and the internal state.

There is a bit of complication stemming from the fact that we pass
paths like "/a", "/b/c/d", while in reality they are more like
"/tmp/filewatcher-XXXXXX/a" or "/tmp/filewatcher-XXXXXX/b/c/d", so we
need to fix our expected results to match the reality.

Another complication there is to know how many times we need to
iterate the watcher before checking the results. Each iteration means
handling one event that comes from the notify watcher. So for each
file system operation we do in the "steps" phase, we need to calculate
how many events we expect.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>