Refactor logging statements in builder-depot crate
This is some of the oldest code we've got and it's about time we take
a pass and refresh the logging output. All logs that you want to wake
you up in the middle of the night are set to `error`. Anything you want
an hourly aggregate on are set to `warn`. Anything that's interesting
is set to `info`. All else is `debug` or `trace` for network messages

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>