Rebuild census if any new rumors are seen
Previously, in an attempt to aggressively optimize things, we would
only rebuild portions of the census according to the type of rumors we
received since the last time through the supervisor's processing
loop. Unfortunately, since the census is dependent on all the data in
the rumors, we could get into strange situations where part of the
census was updated, but was inconsistent with other parts, as seen in

Now, we eschew this optimization in favor of rebuilding the census
whenever we get *any* new rumors, regardless of type. This should
yield a more consistent census overall.

Furthermore, all logic around update counters has now been
consolidated in one place, making it easier to reason about these
operations and facilitate future refactorings.

Signed-off-by: Christopher Maier <cmaier@chef.io>