Fix GH Hook timeouts and auto-build with a forked repo
It is impossible to guarantee that we can search a repository for
all of it's Plan Files so we must pivot to a mechanism where we
best guess where a plan file is and then ask the user to provide
a configuration file (`.bldr.toml`) which provides us a mapping of
plan file locations and any additional files/directories which, when
modified, will cause an auto-build for a project.

With this commit we no longer search the contents of a repository to
find all plan files and then attempt to read them in. We will now
look for a `.bldr.toml` and if one isn't found fallback to a default.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>