Fix the --no-color flag, add a --json flag
Currently this covers the following:
- all launcher output (uses raw arg processing, should this use clap?)
- all hap-sup run output

--no-color and --json are only allowed for `hab sup run` (--no-color
was on all hab sup previously, which let colour leak in from other
components which were called)

My thinking is that this is only really for sending actual sup output
to a logging system / file. Happy to change if needed.

no-color does two things (probably up for debate):
- no color in preamble etc..
- strip all color from output (could be hab messages, could be service output)

The JSON output currently has these fields:
- content
- logkey
- preamble
- file (verbose only)
- line (verbose only)
- column (verbose only)

Potentially preamble could be split to service and group?

Signed-off-by: James Sewell <james.sewell@gmail.com>