Require live members in order for binds to be (initially) satisfied
Previously, a bind was considered satisfied if the bound service group
was merely present in the census. However, in the current rumor
implementation, this group can have no members in it that are
considered alive (as can happen when a member has a long `init` hook
run, for instance).

This can cause confusing (and timing dependent!) behavior in binding
services, particularly if they have templated content that uses the
`eachAlive` template helper.

This commit changes the definition of a "satisfied bind" to
additionally require that there be at least one "alive" member of the
bound service group. This is really what users want when they specify
a binding contract anyway (you aren't really interested in interacting
with dead or uninitialized, services, are you?).

Note, however, that this change really only affects the *initial*
binding behavior when a service starts up. If a bound service group
were to later enter a state where it had no "alive" members, no
additional special processing occurs in the binding service. This
scenario is distinct from initial startup, however. You may have
suspect members due to transient network splits, and altering the
behavior of templating in such cases may have unintended side-effects
(e.g., unnecessary configuration churn, with attendant possible
service interruptions, because of a passing temporary inability to
route network traffic). We will adopt a "wait-and-see" stance on this,
and address it if a genuine need arises.

Signed-off-by: Christopher Maier <cmaier@chef.io>