pkg-export-kubernetes: New component
This generates a simple Kubernetes manifest to standard output. For
now, only the name of the service, the Docker image name and the
number of desired instances are configurable.

In future more things could be modified - topology (together with the
check if number of desired instances is at least 3 for the leader
election to succeed), namespacing, bindings, ring encryption, more
comments in the generated manifest, multiple services in a single
manifest.

Signed-off-by: Lili Cosic <lili@kinvolk.io>
Signed-off-by: Zeeshan Ali <zeeshan@kinvolk.io>
Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>