hab: Hook pkg-export-kubernetes to hab CLI
The kubernetes exporter will be available under `hab pkg export
kubernetes`. It is done in the same way as with the docker exporter,
which is a bit hacky, because calling this with some command alias
won't work (`hab package export kubernetes`).

We should also consider adding a shortcut like `hab pkg export k8s`.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>