Fix issue configuring worker bldr_url when first bind member is dead
`.first` doesn't guarantee it will return an alive member. Only alive
members have gossip configuration applied in a service group. Using
the `.first` helper with a bind can result in a stale value.

This commit ensures that we find an alive member to pull our depot
configuration from for the worker.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>