Add initial_file_event arg to FileWatcher::run
The additional parameter serves to control whether an event should be
emitted if, at the time the FileWatcher starts, the watched file is
already present.

In the case of the PeerWatcher, this is exactly the behaviour we want,
because in Kubernetes, when a pod starts its volumes (and the peer file
among them) have already been mounted.

Signed-off-by: Lorenzo Manacorda <lorenzo@kinvolk.io>