[studio] Install the same `a.b.c` version of backline as Studio.
This change ensures that the same `a.b.c` version of the Studio's
backline package gets installed as the version of Studio that's running.
In other words, if you are running
`core/hab-studio/0.78.0/20171015010203`, then the version of backline
that will be installed would be `core/hab-backline/0.78.0`.

New Environment Variable
------------------------

While working on this change, I found an environment that I left in the
Studio code to override the package identifier default for backline.
This ended up being really useful for working on the solution, but the
naming of the environment variable was less consistent with the set of
variables we have now that relate to the Studio. So, I renamed this
variable to `HAB_STUDIO_BACKLINE_PKG`.

Closes #3712

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>