Make service config code easier to test
Introduce the PackageConfigPaths trait, which will allow us to
override the paths where the user configuration should be put for
testing purposes. Before that, the paths were basically hardcoded to
something under `/hab`. There is an environment variable
(`TESTING_FS_ROOT`) for changing the filesystem root, where `hab`
directory is being placed, but it is not really usable from unit
tests.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>