[airlock] Use clap's TrailingVarArg setting to accept cmd args.
This change drastically simplifies the CLI parsing when interpreting
where the clap CLI parsing should terminate. The `TrailingVarArg`
setting can be used on commands and subcommands where the last
positional argument has `+multiple` set. The only drawback is that there
is not a differentiation between the cmd (i.e. `ARGV[0]`) and the
program arguments. However, given the benefits, this is more than worth
the price of a slightly less clear CLI usage message.

Note to self: this knowledge can be backported to the `hab pkg exec`
subcommand which has the same clap CLI parsing issue.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>