[builder-worker] Allow for optional Airlock & Airlock networking.
This change adds the ability to disable Airlock networking and also to
disable Airlock entirely, falling back to an uncontained Studio
execution. Both of these modes are intended to aid in development,
potentially execution in single tenant environments, or environments
that cannot properly support Airlock. The defaults for production are to
enable both Airlock and Airlock networking.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>