plan-build changes to implement environments
To specify runtime and buildtime environment values in a predictable
way, we provide several helper functions, as well as a new callback,
`do_setup_environment`.

We do this for a few reasons, versus just allowing users to perform
raw environment manipulation in their plans.

First, we need to ensure that the runtime environments of any
dependencies are properly layered. This includes extending any
`PATH`-like variables and preventing the unintentional clobbering of
any existing values that may be present.

Second, some environment variables (notably `PATH`, but also things
like `CFLAGS`, LDFLAGS`, and the like) are actually determined by the
values of other Habitat plan variables (PATH by `pkg_bin_dirs`,
`CFLAGS` by `pkg_include_dirs`, etc.). By controling the environment
creation process, we can better ensure the values of these variables
are what they should be.

Third, we want to ensure that all environment manipulations take place
in the appropriate order. Thus, we resolve all environment information
from dependencies first, then add any additional variables that the
plan author has specified.

(It is true that this is all "just Bash", and we thus cannot 100%
guarantee that a determined user has not modified things contrary to
this design. However, we want to make it as easy as possible for users
to adhere to this design, and as difficult as possible to do
otherwise.)

The provided functions are:

* set_runtime_env
* push_runtime_env
* set_buildtime_env
* push_buildtime_env

The construction of the various environments is described below.

The "runtime environment" is initially empty. Then, the direct
dependencies of the package are consulted, in `pkg_deps` order, to
pull in any runtime environment values they may have
specified. `PATH`-style variables (or "aggregates" --- those that have
multiple delimited values) are layered appropriately, with newer
values pushed to the front.

Thus, for example, with a `pkg_deps` list of
`(myorigin/foo, myorigin/bar, myorigin/baz)`, the resulting value for
the `PATH` environment variable might look like this:

    PATH="${PATH_FOR_BAZ}:${PATH_FOR_BAR}:${PATH_FOR_FOO}"

Stated another way, "last one wins".

Non-aggregate values are overwritten in the same "last one wins"
manner.

Finally, values specified by the user using `set_runtime_env` and
`push_runtime_env` in their `do_setup_environment` callback are
applied to the assembled runtime environment.

The "buildtime environment" is assembled in a similar way, except it
is sourced from `pkg_build_deps` instead of `pkg_deps`.

Finally, after all user modifications have been applied, these
separate environments are applied to the system environment in the
build studio. First, the runtime environment is applied on top of the
system environment, followed by application of the buildtime
environment on top of that.

`RUNTIME_ENVIRONMENT` and `BUILDTIME_ENVIRONMENT` metadata files are
generated based on the computed values of those separate environments,
and are not pulled from the final system environment that the build
ran in. This prevents them from becoming cross-contaminated, which is
important, since the value for `RUNTIME_ENVIRONMENT` is used as input
to dependent builds (`BUILDTIME_ENVIRONMENT` exists purely for
documentation purposes, and is not actively consumed by any other
process).

In addition to these, `RUNTIME_ENVIRONMENT_PROVENANCE` and
`BUILDTIME_ENVIRONMENT_PROVENANCE` metadata files are also
created. Like `BUILDTIME_ENVIRONMENT`, these are "passive" files that
are not consumed by any other process. They show which packages the
values for individual environment variables comes from, which can help
debug builds using multiple levels of environment inheritance.

Finally, this fixes errors in generating the `PATH` environment
variable.

Previously, we were including the entire *build time* path in a file
that is intended for *run time* consumption. This included paths to
build dependencies, as well as blank entris and system paths from
within the Studio itself.

Moving forward, the `PATH` metadata file will contain the complete
path of a package, including all directories specified in
`pkg_bin_dirs`, as well as those `pkg_bin_dirs` of all its direct runtime
dependencies (whose paths include *their* direct runtime dependencies,
etc.). No build dependencies will be present.

The `RUNTIME_ENVIRONMENT` metadata file will also contain an entry for
the `PATH` environment variable, and its value will be identical with
the contents of the `PATH` metadata file. The Supervisor will
preferentially pull a value for `PATH` from `RUNTIME_ENVIRONMENT`; the
`PATH` metadata file is provided for backward compatibility, and may
be removed in the future.

Note that the content of the `$PATH` environment variable presented to
the Supervisor is controlled completely by Habitat; you are not
prevented from export values for `$PATH` in your plan, but only the
directories detailed above will make it into the `RUNTIME_ENVIRONMENT`
and `PATH` metadata files. This helps preserve the self-contained
nature of Habitat packages; the value of `$PATH` seen by the
Supervisor will only contain paths that are guaranteed to be present
on the system.

When layering environment values from dependencies, existing `PATH`
metadata files are used when `RUNTIME_ENVIRONMENT` files are not
present (as will be the case for every Habitat package made prior to
the release of the current feature!). In the near term, these paths
will continue to contain the extra and anomalous entries described
above. However, as lower-level dependencies are rebuilt with proper
paths, these proper paths will cascade outward with successive
rebuilds; eventually, all paths should be properly constructed.

Signed-off-by: Christopher Maier <cmaier@chef.io>