Pull PATH from RUNTIME_ENVIRONMENT preferentially
When we need access to a `PATH` from a package, we now preferentially
pull it from the `RUNTIME_ENVIRONMENT` metadata file. If no such file
exists, we fall back to the older `PATH` file (the previous behavior).

Extracted an `existing_metafile` function which just tests for the
existence of a metadata file to aid in efficiently implementing this
fallback logic.

Signed-off-by: Christopher Maier <cmaier@chef.io>