pkg-export-kubernetes: Proper error handling
We were using String as our error, which is fine for the first
implementation but in the following patch, it won't suffice when we add
docker image generation code.

This duplicates some error handling code with the docker crate but we will
remove that when we start using docker crate.

Signed-off-by: Zeeshan Ali <zeeshan@kinvolk.io>