Fix paths inconsistencies errors
There were two bugs:
- repeated file disappeared events for the same file
- repeated unwatch actions for the same file

This happened when deleting a directory with watched files quickly, so
we were getting a batch of RemoveNotice events, which after 2 seconds
would be followed by a batch of Remove events.

The handling of RemoveNotice event involves traversing the chain
starting from the removed file until the end and generating the drop
watch action for every traversed item. This happens in the drop_common
function. So if we quickly removed a file and its parent directory
then drop_common would generate a drop watch event for the file and
then two drop watch events for the parent directory and the file
again.

Also, every time the function checks the real_file member to check if
we should generate the file disappearance action. So here is the
similar problem like with drop action generation above.

Since the chain is like a linked list, unwatching some item in the
middle of the chain will certainly make the real file to disappear. So
set the real_file to None every time an item is going to be
unwatched. Also, set the next member of the previous item in the chain
to None for every unwatched item, so we don't generate repeated drop
watch actions for the following NoticeRemove events.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>