Ensure permissions are set on the config directory
NOTE: This is a temporary bandaid for the micro-outages we've been
having with the API server every time the service restarts. The real
fix lies in how the Supervisor starts services (and more specifically,
which user they are started by). Addressing that should (partially)
unlock new behaviors for the Supervisor (like the ability to more
sanely run as a non-root user).

Fixes: #3521

Signed-off-by: Christopher Maier <cmaier@chef.io>