Update build dependencies for pkg-export-docker
The code from #4041 (and later, #4201) introduced `backtrace-sys` as a
dependency for the `pkg-export-docker` crate. This needs `make` as a
build dependency, which we add here to the Habitat plans for both
`core/hab-pkg-export-docker` and
`core/hab-pkg-export-kubernetes` (which has a library dependency on
the `pkg-export-docker` crate)

This was uncovered only in Travis, likely because most development on
Habitat is done on Linux systems, and thus using the default Linux
chroot-based studios. This ends up reusing compiled Rust artifacts in
`/src/target`, which will have included the `backtrace-sys` artifact
compiled by the Linux host system. As a result, package builds in
those studios ran without an issue. Running the build in a
Docker-based studio (e.g., on MacOS) immediately revealed the issue.

Signed-off-by: Christopher Maier <cmaier@chef.io>