Fold building logic into deploy script
Previously, we would consult an `AFFECTED_DIRS` environment variable
to see if we needed to build packages. This would trigger builds for
every package if any code changed for any packages (e.g., we'd build
everything even if we just added a comment to the `plan.sh` file of an
exporter).

Now, we eschew the `fast_pass.sh` script for deploy steps, instead
driving the decision on whether to build a package from the contents
of `.bldr.toml` and the state of the Git repository, using the new
`should_we_build.sh` script.

Packages are still built serially according to the order laid out in
the `COMPONENTS` environment variable, since we don't yet have a
better way to convey the dependency chain between packages (since many
have implicit dependencies).

Signed-off-by: Christopher Maier <cmaier@chef.io>