pkg-export-docker: Method to add mandatory pkg argument
I previously kept the mandatory package argument out of the crate because
the only user of this crate (kubernetes exporter) doesn't allow user to
specify multiple packages at the same time. I now realized that the
'multiple' nature of this argument can just be made configurable through a
boolean argument.

Signed-off-by: Zeeshan Ali <zeeshan@kinvolk.io>