pkg-export-k8s: Parameterize Manifest::write
The Manifest shouldn't be creating and keeping the 'Write' instance itself
but rather its 'export' method should get it as a parameter. This gives
users of this crate the ability to decide where to write the manifest to,
and makes it possible to easily implement Clone trait on the Manifest.

Signed-off-by: Zeeshan Ali <zeeshan@kinvolk.io>