Be more thorough in generating PATHs from older packages
With the addition of the new runtime environment variables feature, it
has become apparent that our fallback logic for extracting `PATH`
variables from older packages built before version 0.50.3 was not as
complete as it needed to be.

Now, when we encounter a dependency that does not have a
`RUNTIME_ENVIRONMENT` metadata file, we not only take the `PATH`
metadata from that dependency, but also from its _transitive_
dependencies.

This should more fully replicate the older behavior of the Supervisor,
which assembled this information at runtime prior to the 0.50.3
release.

Signed-off-by: Christopher Maier <cmaier@chef.io>