Specify group ID for service group in exported containers
Previously, we allowed users with specific needs to customize the user
ID of the service user when exporting a container. Now, we also allow
customization of the group ID of the service group as well, instead of
having it always be the same as the user ID.

The short option flag for `--user-id` was also removed (and none was
added for `--group-id`); these are special-use options, and should
only be needed in special cases. There's no reason to claim valuable
short flags for this.

Signed-off-by: Christopher Maier <cmaier@chef.io>