Remove all embedded color codes
In several places we were directly embedding ANSI color codes in a log
message as a way to emphasize certain parts of that message. However,
this bypasses the existing controls we have over whether or not to
colorize the Supervisor's output; regardless of your settings, these
emphasized parts would *always* be colored.

As an initial step to bring colored output under control, we are
simply removing all these embedded "emphasis colors". A possible
long-term solution would be to introduce a minimal templating language
of sorts that allows us to mark particular bits of text for emphasis
of one kind or another, allowing the final formatting of a message to
be handled centrally. Think of an HTML/CSS-style separation of
concerns, but for fancy colored log output.

Signed-off-by: Christopher Maier <cmaier@chef.io>