Launcher honors `--no-color`
The `core::output` colorization logic hinges on the value of a global
atomic variable. However, the Launcher is a separate process from the
Supervisor; any value set in the Supervisor is completely distinct
from what the Launcher sees. As a result, the Launcher would *always*
colorize output, because colorization currently defaults to `true`.

Since the Launcher always has access to the arguments used to start
the Supervisor, and since there is only a single option we care about
here, it's a simple matter to just check if a `--no-color` argument
was passed and then take action accordingly. There's no need to bring
`clap` in for something like this.

Signed-off-by: Christopher Maier <cmaier@chef.io>