Remove debug() implementation; log `try()` messages to STDERR
By emitting messages to `STDOUT`, the original implementation would
break anything that tried to capture the output of a call to `try`,
which we do in one place.

Furthermore, we already have a debug mechanism (based off the `$DEBUG`
environment variable, and using `set -x`). More refined log levels can
be work for the future.

Fixes #4380

Signed-off-by: Christopher Maier <cmaier@chef.io>