CI: Update rustfmt to latest nightly: 0.3.4
Version 0.9.0 is already pretty old, considering that rustfmt tool
improves very fast. This will mean having to change formatting in a lot of
files but the sooner we update the lesser the needed changes. We don't
need to fix formatting of all files right away but leave that as a
homework for contributors to fix formatting in the files they touch with
their PR.

Signed-off-by: Zeeshan Ali <zeeshan@kinvolk.io>