Use the new network abstraction in butterfly client
This commit makes the `Client` type to be parameterized with a type
implementing the `GossipSender` trait (in rust, this changes the type
from `struct Client` to `struct Client<GS: GossipSender>`).

This also changes the prototype of the `Client::new` function, so now
it takes a socket instead of creating it and thus it is not possible
for this function to fail anymore.

The tests are hardcoded to use the `RealNetwork` and
`GossipZmqSocket`, but to avoid repetitive code, some helper functions
for creating a `Client` instance were added.

This change also affects the "hab config" utility, so it also uses the
`RealNetwork` and `GossipZmqSocket`.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>