Simplify port handling in butterfly server
The butterfly server allowed to specify a separate port number for
swim connections and a separate port for gossip connections. The only
serious user of the butterfly server, the supervisor, was passing the
same number in for both. It works, because even if butterfly is still
opening two ports, one is a UDP port and the other is TCP port.

Another reason for this change is to simplify the discovery of the
exposed address and port for the supervisor being behind a NAT. This
feature may come in future work revolving around the spanning ring.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>