Add network abstraction for butterfly
This commit adds a couple of traits as an abstraction for
networking. The abstraction is not trying to be very generic - it is
tailored for the butterfly server and client needs, so it may be a bit
leaky (for example, would be nice to have address and port under a
single abstraction like Address instead of having them separate as
Address and AddressAndPort, but it has to be that way, because we are
sending ports in protocol messages), but it's fine for its intended
purposes.

The reason for introducing the abstraction is to be able to use a
different implementation for testing needs. For example to mock a
scenario which may be hard to simulate with UDP and ZMQ sockets on
localhost. An example of the scenario would be talking to the
butterfly server behind a NAT.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>