Remove HAB_NON_ROOT environment variable
This was put in place to allow packages to be installed as a non-root
user. It doesn't really provide much in the way of protection, though,
since the user in question would still need write permissions on the
`/hab` directory (and various subdirectories) for the operation to
succeed; it's more of an advisory feature, allowing the user to say
"yes, I know; just let me do what I want to do".

Signed-off-by: Christopher Maier <cmaier@chef.io>