Add builder-base-plan back
This file was initially removed when Builder itself was removed from
this repository. It contains common plan callback implementations used
by many of our Rust-based packages. In particular, it is required by
`eventsrv` and `launcher`, two packages that remain in this
repository.

Adding the file back allows us to build these packages again. Longer
term, however, the rest of the packages in this repository should take
advantage of it (many of them duplicate this code unnecessarily), and
functions and variables should be renamed to reflect the fact that
they are not Builder-specific.

Signed-off-by: Christopher Maier <cmaier@chef.io>