consistently use softwrapping on paragraphs
I touched these two last night and noticed there is an inconsistent mix
of softwrap and various lengths of hardwrap. I let it go then but after
a fresh sleep I still couldn't let it stand. I've moved it to softwrap
all around because it's far easier to edit if you don't have to
manually rebalance thing.

This is only updating the things I touched but I aqm happy to to this to
the REST of the file if useful/desired.

Signed-off-by: David Aronsohn <WagThatTail@Me.com>