`hab svc load` respects currently-installed software
Previously, running `hab svc load` would unconditionally reach out to
Builder to install a package. This can cause issues if you're working
with packages locally that are not present in Builder, or if you're on
the Habitat team installing new software during a maintenance window
when Builder is down.

That behavior is also at odds with that of `hab svc start`, which
starts locally-installed software if present, before going out to
Builder as a last resort. It's also at odds with `hab svc load`
itself, when you're updating an existing service!

This PR just uses the existing code in the `hab svc load` path. It
should have done that from the beginning; the fact that it didn't was
an unfortunate oversight.

Fixes #4019
Fixes #4065
Fixes #4602

Signed-off-by: Christopher Maier <cmaier@chef.io>