Raise the limit of open file descriptors to the hard limit
The network abstraction caused the threads to hold on a bit more open
file descriptors - RealNetwork holds one and Inbound now holds two
instead of one (one as a sender and one as a receiver). The threads
are detached and nothing explicitly terminates them, so they are
holding the descriptor until the process is terminated. This together
with the fact that cargo builds one big binary with all the tests
where many of them spawn multiple server which in turn spawns many
threads led to the file descriptor exhaustion.

This workaround raises the soft limit to the hard limit (which means
going from 1024 to 4096 on my computer), so the limit won't be hit too
easily.

The proper fix I believe would be rather to tell the threads to shut
themselves down when we want to terminate the server. But this is
likely a lot of work…

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>