Add network switchboard
Network switchboard is a channel based emulation of a NATed
network.

The switchboard can be used to generate zones (which simulate private
networks) and establish a relationship between them. Doing it results
in getting a NAT object.

NAT object can be used to punch holes in it. The holes are basically
addresses we can use to pass to the butterfly servers as additional
addresses.

The butterfly servers are created using switchboard.

After punching some holes and creating the servers, we need to make
sure that we create a route between the punched hole and the internal
address of the server.

After this layouting is done, we can tell the servers to start talking
to other servers.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>