Complete documentation coverage of all templating variables
Adds coverage for `srv` and `bind` keys, as well as typing and
examples for all content.

Added notes on all data that ultimately makes its way to the rendering
context, as a warning against changing them. Long term, we'll need to
explicitly decouple the data in the templating context from our
internal implementation details.

Fixes #4626

Signed-off-by: Christopher Maier <cmaier@chef.io>