[studio] Update stage1 tarball.
This change updates the `stage1` Studio type implementation to use a
newer stage1 tarball which has updated software and version of the `hab`
CLI.

As part of the tarball updates, a change was made to the Linux From
Scratch instructions (on which the stage1 tarball is based) which drops
a file creation step that the Studio code performed. This change will
still run the old behavior if that file exists and skips the step when
the file doesn't exist. A future commit may remove this check altogether
as a final clean task.

Finally, the URL scheme is updated from `https://` to `http://` as the
`wget` implementation that BusyBox provides isn't SSL-aware out of the
box.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>