[common] De-emphasize "channelness" in the install logic.
This is an internal refactoring of the install logic to introduce a
simple type representing a release channel. Future work could and should
expand on this type and ultimately extract it into the `core` crate for
more general use.

The goal with this change was to reduce the channel-obsession present in
this module and attempt to keep work related with channels scoped to the
activity of fetching the latest package version and to display channel
recommendations when a package was not found the requested channel. The
internal `InstallTask` type now takes a non-optional `Channel` allowing
the rest of the code to count on a set channel.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>