[common] Add `FullyQualifiedPackageIdent` to install logic.
This change is an internal refactoring of the install logic which
introduces a new type called `FullyQualifiedPackageIdent` which
represents...a fully qualified package identifier. This allows any code
which requires such an identifier to use a type which guarantees these
promises.

The type uses an internal `Cow` of a `PackageIdent` which allows the
user of this type to hold borrowed and owned idents without additional
cloning. This type helped to update several functions and methods with
clearer expectations and led to less `.is_some()`-style checking.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>