[common] Explicitly report latest version checking in install logic.
This change adds back channel information back to a user, but in a more
correct location. Prior to this change (and a previous refactoring), the
install logic would report a line like the following:

```
Installing core/redis from channel 'stable'
```

However, no installing would necessarily happen from the stable channel.
Rather, we check for the latest version of a package scoped to a channel
and use that information to determine next steps.

This change adds a status line specifically mapping to the Builder API
call where we check for the latest version and get back the fully
qualified package identifier in a JSON response.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>