[common] Canonicalize paths when copying artifact to cache.
This change is intended to close any remaining edge conditions occurring
when a user tries to install a `.hart` file directly from the artifact
cache directory. For example:

```
hab pkg install /hab/cache/artifacts/core-redis-*.hart
```

Without a sufficient guard, the install logic will effectively try to do
the following:

```
cp /hab/cache/artifacts/artifact.hart /hab/cache/artifacts/artifact.hart
```

That is, the source file will be copied over and onto itself, usually
resulting in a zero-sized file--obviously not what was expected.

There is currently a check for this in the install logic, however both
the source and target paths are not canonicalized (dereferencing
symlinks, relative paths, etc.). This change fixes that and adds a debug
line in both the `true` and `false` arms of the check for debugging
completeness.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>